requirements.txt
Django==5.1.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-environ==0.11.2
django-cors-headers==4.4.0
drf-spectacular==0.27.2
psycopg[binary]==3.2.3
celery==5.4.0
redis==5.0.7
python-telegram-bot==21.6
requests==2.32.3
PyJWT==2.9.0

# Dev / QA
pytest==8.3.3
pytest-django==4.9.0
flake8==7.1.1
coverage==7.6.4